+++ src/main/java/org/miles2run/business/services/redis/CommunityRunRedisService.java
@@ -104,7 +104,7 @@ public class CommunityRunRedisService {
                 Response<Long> citiesCountResponse = pipeline.scard(String.format("%s-cities", slug));
                 Response<Long> runnersCountResponse = pipeline.scard(String.format("%s-all-runners", slug));
                 pipeline.sync();
-                Long totalDistance = totalDistanceCoveredResponse.get() == null ? Long.valueOf(0L) : Long.valueOf(totalDistanceCoveredResponse.get());
+                Double totalDistance = totalDistanceCoveredResponse.get() == null ? Double.valueOf(0) : Double.valueOf(totalDistanceCoveredResponse.get());
                 Long totalDuration = totalDurationResponse.get() == null ? Long.valueOf(0L) : Long.valueOf(totalDurationResponse.get());
                 return new CommunityRunStats(runnersCountResponse.get(), countriesCountResponse.get(), citiesCountResponse.get(), totalDistance, totalDuration);
             }
@@ -125,7 +125,7 @@ public class CommunityRunRedisService {
         return jedisExecutionService.execute(new JedisOperation<Boolean>() {
             @Override
             public Boolean perform(Jedis jedis) {
-                return jedis.sismember(RedisKeyNames.COMMUNITY_RUNS,slug);
+                return jedis.sismember(RedisKeyNames.COMMUNITY_RUNS, slug);
             }
         });
